テーマ関連の設定を別ファイルに
